Add html attributes to label elements
This adds an addition optional parameter to the following methods:
- email_field
- password_field
- number_field
- phone_field
- range_field
- search_field
- telephone_field
- text_area
- text_field
- url_field

The  original use case was to just add custom css classes to the
label element but I think a better approach would be to allow any
other html attribute to be added.

Usage eg.

= form_for person do |f|
  - f.text_field :name, {label_options:{
                          class: 'my-css-styles',
                          style: 'color:red;'
                         }}

label_options hash is optional and accept any html additional html
attribute you would like to use in the label.