Allow value and text method options to radio_button_fieldset (#82)
When generating a radio-button fieldset, there are times when we want to specify
how to get the form value or label text for the choices passed in, the same way
`collection_radio_buttons` does in Rails. For example, when assigning a team to
a person we may want to use the team's DB ID as the value and the team's name as
the radio button text. This change adds these options to
`radio_button_fieldset` (adding them to other places, such as
`check_box_fieldset`, is an exercise for later).