Accessibility fixes
Based on GOV.UK design system, and constrained by the fact that this gem is old and not using the latest version of the UI components, we've introduced a couple minor improvements:

* Replace the role `group` with role `alert` in the `div#error-summary`.
* Change `h1` to `h2` in the heading of the error summary.

Also, we've bumped some gem dependencies to fix most of the vuln alerts, but in a conservative way as this gem is suppose to be used in the context of a full app that will be in control of actually specifying the versions of the dependencies they want to use. I don't want to "force" specific versions in this gem.

A new version of the gem will be built and published with these changes.