Update error summary to new design system
This will only update the error summary to output the correct markup
and styles for the new design system. It will not update the rest of
the components as that is still pretty much WIP.

It will also removes the hard dependencies with the gems
`govuk_frontend_toolkit` and `govuk_elements_rails`.

This commit is not intended to be merged. Just wanted to have it as
a reference if needed in the future.