Add ability to override the legend text
There are some situations where the legend text can't just be picked from locales, for example because it has dynamic content. We need this in one of our services.

With this small change we provide the ability to pass the content of the legend text.

It uses the same existing `legend_options` hash and adds a new `text` option.
It is completely backwards compatible and the default behaviour continues to be to pick the legend text from i18n locales.